Fix a few of the cross platform compatibility problems in wvstreams/utils
git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@941 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ utils/wvtclstring.cc
@@ -106,7 +106,7 @@ WvFastString wvtcl_unescape(WvStringParm s)
     
     const char *cptr, *end = s + slen - skipquotes;
     char *optr = out.edit();
-    for (cptr = s + skipquotes; cptr < end; cptr++)
+    for (cptr = s + (int)skipquotes; cptr < end; cptr++)
     {
 	if (*cptr == '\\')
 	{
